batch_interface: remove `rounds` data member
`rounds` was used to track the number of times the scratch space was
cleared by batch_context_add_* for transparent verification. It will
also tell the number of terms the user entered (rounds*capacity + len).

Keeping track of this data was not worth it since the user could always
have a counter to for total number of terms entered. It also made the
`batch_verify` function look messy. Hence, decided to remove `rounds` memeber.

The side effect of removing this data memeber is setting `batch_ctx->result = 1`
during creation since we alwasy need to `&&` with `batch_ctx->result` during
verification.